PythonProj/Final_Pro/requirements.txt
asn1crypto==0.24.0
astroid==2.4.2
attrs==17.4.0
Automat==0.6.0
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==20.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coverage==5.3
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
isort==5.5.2
Jinja2==2.10
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.0
mccabe==0.6.1
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.6.0
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pylint==2.6.0
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.3
python-debian==0.1.32
python-string-utils==1.0.0
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.15.0
ssh-import-id==5.7
systemd-python==234
toml==0.10.1
Twisted==17.9.0
typed-ast==1.4.1
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
virtualenv==15.1.0
wrapt==1.12.1
zope.interface==4.3.2
